keyboard: Allow passing in a `networkx.DiGraph` (directed graph) instance
This gives us more flexibility in how we specify the state machine. For
example you might write your own helper function that generates a
DiGraph instance, which might be easier than generating a string.

I still like the string format because it's simple, and the way you
have to call `networkx.parse_edgelist` is non-trivial, so I've kept it.

Note that `MultiDiGraph` is a subclass of `DiGraph`, but it won't
actually work because several networkx functions behave
differently (return a different data structure) for multi-graphs (which
are graphs with multiple parallel edges, i.e. several edges with the
same start & end nodes but perhaps different attributes).