Keyboard.enter_text: Fix navigation after the first letter
It was always calculating the path as if it was starting again from the
initial position (before it had entered any of the previous letters).

Sometimes this would work, because `navigate_to` has a loop where it
double-checks that it landed where it expected, and then recalculates.
Other times it might end up navigating away from the keyboard (e.g. into
the search results).

This bug was introduced in commit 5756985 (keyboard: Move `page`
parameter from constructor to `enter_text`). I suppose I didn't re-test
against a real set-top-box after doing that change.

This bug wasn't caught by the unit tests because it was using a class
that wasn't a real `stbt.FrameObject` -- in particular it didn't have
the immutable behaviour of FrameObjects, so the `selection` property was
updating even when the `page` instance wasn't properly refreshed. Now
I have split the test's implementation into 2 classes:

* `YouTubeKeyboard` is the "device under test" -- it knows what the
  actual state is at all times.
* `_Keyboard` is a real `stbt.FrameObject` subclass, just like you'd use
  in your test scripts. The only difference with a real test script is
  that it overrides `__init__` and `refresh` so that it can wire in our
  fake device under test.