pytest.ini: Configure log_format
Pytest already captures & shows logging calls, similar to how it
captures stdout/stderr. It shows each of these in separate sections:
"Captured stdout call", "Capturred stderr call", and "Captured log
call", so it helps if the tests use logging (rather than print) to see
the log messages interleaved in the right order.

Pytest's default log_format is the same as this, but without the
timestamp. Pytest's default time format is %H:%M:%S which should be
sufficient for unit tests; we don't really need the date or the
timezone.

Note that this captures all log levels, including DEBUG. If we set
"log_level=DEBUG" in pytest.ini, then it *also* logs those calls to
stderr and we end up with duplicate logs.