_stbt.logging: Integrate structlog with stdlib logging
So we can log (internally) using [structlog], but it still gets printed
by the stdlib `logging` framework, so that users can configure the
output in the same way that they configure their own logging. We expect
that users will use `logging` directly, not `structlog`.

In future we can "tee" the output from structlog to a structured
file (e.g. jsonl) while continuing to print it to stderr.log for human
users' consumption.

[structlog]: https://www.structlog.org/