stbt.Region: Disallow calling `intersect` on an instance
`intersect` is a `@staticmethod`, but you can still call it on an
instance -- and the instance will be ignored, not considered among the
arguments.

So `r1.intersect(r2)` behaves like `stbt.Region.intersect(r2)`, not like
`stbt.Region.intersect(r1, r2)`.

Before v30, calling intersect with a single region wouldn't have been
allowed because it required exactly 2 arguments; but since v30 (when we
allowed an arbitrary number of arguments) this incorrect usage wouldn't
cause an assertion and even stbt lint wouldn't catch it.

Similarly for `bounding_box`.