Region: Allow translating one region by another one
...in fact this will allow translating by any tuple with x and y as the
first two items.  This is helpful when dealing with items that have
substructure.  For example: A carousel might have multiple items laid out
in a grid:

    +---------------+     +----------------+
    |    Title 1    |     |    Title 2     |
    |               |     |                |
    | Description 1 |     | Description 2  |
    | this TV show  |     | this one isn't |
    | is really     |     | so interesting |
    | good and      |     |                |
    | exciting      |     |                |
    |               |     |                |
    +---------------+     +----------------+

The boxes might be given by `REGION[0]` and `REGION[1]`.  You could then
have TITLE_REGION and DESCRIPTION_REGION relative to the boxes.  To read
all the information it might look like:

    for r in REGION:
        title = stbt.ocr(region=TITLE_REGION.translate(r))
        description = stbt.ocr(region=DESCRIPTION_REGION.translate(r))