stbt.Region: Document static methods in `Region` docstring
Since the previous commit, `bounding_box` and `intersect` are defined in
a MetaClass. I can't get sphinx to show them, even with
`:inherited-members:`.

I have also had to move the `width` and `height` properties and the
`from_extents` factory into the docstring, to get the right order in the
generated documentation.