Replace treating `None` as `Region` with `Region.EMPTY`
This may make it easier to use functions that return a `Region` that might
be `None` because in many cases they may not need to check `region is None`
before performing operations on that region.  It also makes it easier to
implement functions with take a default `Region` because we can follow the
standard `region is not None` to mean "did the caller pass a region
argument".  For an example of this see `Grid`.