stbt.Keyboard: Verify that you've reached the target
Previously the only verification was in the while loop's condition:
`while current != target`. So if your model of the keyboard was wrong,
`navigate_to` would keep on trying to reach the target until
`navigate_timeout` was reached (default 20 seconds). Maybe this would
eventually reach the key, working around gaps in your model; but in
reality it leads to very confusing, hard to debug, behaviour.

Now we verify that we've reached the target letter after each sequence
of keypresses, so we fail earlier and we provide a useful error message.

What *isn't* a goal of the original implementation: Work around missed
keypresses or double keypresses. With a correct lirc configuration you
should never see missed/double keypresses.