stbt.Keyboard: Add `verify_every_keypress` for debugging
Since the previous commit, we verify that we've landed on the target key
after the first attempt. For example on a qwerty keyboard to get from Q
to D we press KEY_DOWN, KEY_RIGHT, KEY_RIGHT and we verify after the 3rd
keypress that we've landed on D.

If your keyboard navigation is behaving incorrectly --perhaps because
your model (`graph`) of the keyboard is incomplete-- you can set
`verify_every_keypress=True` to enable additional assertions to help you
debug your model. In the example above, we would verify that KEY_DOWN
landed on A, then that KEY_RIGHT landed on S, and finally that the last
KEY_RIGHT landed on D.

The default behaviour is `verify_every_keypress=False`, which is faster
and closer to how a human navigates an on-screen keyboard: We eyeball
the distance to the target letter and press the arrow buttons a bunch of
times, and then we correct if we're short or we've overshot.