utils: Fix `to_bytes` and `to_unicode` to return real native types
Sometimes they would return python-future's newtypes which behave oddly in
certain circumstances.  For example:

    b"Hi %s" % newbytes(b"There")

would be `"Hi b'there'"`.  This was causing `press()` to fail as we'd ask
`lircd` for key `b'KEY_RIGHT'`.

This commit fixes `to_bytes` to always return a real native `bytes` object
and so fixes the lirc remote issue.

I'm of the opinion that as a library that intends to support Python 2 and 3
`stbt` shouldn't leak that we've implemented this support with
python-future so we should return normal Python types wherever possible.
We need to support both Python 2 and 3, but it doesn't make sense for our
customers' test-packs to be python-version agnostic.

I've added `text_type` and `native_str` to `_stbt.utils` here rather than
`from future.utils import text_type, native_str` becuase the implementation
is *trivial* and I'd like to drop the dependency on python-future when
we're using Python 3 (at some point in the future).  I figure that
`python-future` is supposed to let you "just" write in Python 3, and
convert that to Python 2 as necessary, so it doesn't make a lot of sense
to require it when running on Python 3 natively.