Fix _fake_cec for Python 3
`buffer` is not a builtin on Python 3, it's called `memoryview` instead.
`memoryview` is also present on Python 2.  `.encode` doesn't exist on
`bytes` in Python 3, and you can't encode `unicode` as hex, so we use bytes
here.

In this case there is no `from builtins import *` at the top of the file,
so `bytes` means `bytes`, etc.