Don't expose future.types.newxxx to our users
I've checked all our code for instances of `bytes(` or `str(`.  On Python 2
with the `future` library they can actually return instances of `newbytes`
and `newstr` rather than what you'd expect.  This changes some calls to
`str` to be explicit about what type you'll get, and to avoid these
future newtypes.  future newtypes can behave oddly so I want to avoid
exposing them to our users.

I'm of the opinion that as a library that intends to support Python 2 and 3
`stbt` shouldn't leak that we've implemented this support with
python-future so we should return normal Python types wherever possible.
We need to support both Python 2 and 3, but it doesn't make sense for our
customers' test-packs to be python-version agnostic.