Python-future: Avoid overriding `bytes`, `dict`, `int`, `range`, `str`
...with explicit imports.

I think these imports lull you into a false sense of security.  Lets say
you see

    from builtins import *

    ...

    for k, v in d.items():
        ...

My initial reaction was, "ah it's all fine, we're in Python 3 mode so this
iterates through the items rather than creating a list and iterating
though that".  Unfortunately this isn't true, it depends on the
provenence of `d`.  On Python 2 `d` will be a regular Python 2 dictionary,
with old Python 2 behaviour, if:

* `d` was passed to us from the user
* `d` came from some other Python function like `json.loads()`
* `d` was created as a literal with `{...}`

It only has this new behaviour if it was explicitly created or converted
in our code with `dict(...)`.

The same is not true of the function-like imports such as `filter`.  If
you see `filter` in our code you know it has the Python 3 behaviour, i.e.
it's lazy.

Interestingly it seems that the other future builtins don't return
newtypes. On Python 2:

    >>> type(future.builtins.chr(23456))
    unicode
    >>> type(future.builtins.oct(246))
    str

so that's one concern I had gone.

The full list of builtins that [python future can provide][1] are:

    from builtins import (ascii, bytes, chr, dict, filter, hex, input,
                          int, map, next, oct, open, pow, range, round,
                          str, super, zip)

[1]: https://python-future.org/imports.html