Add `for_object_repository` decorator
A decorator that marks classes and functions so they appear in the Object
Repository.

Classes that directly derive from `stbt.FrameObject` have this decorator
applied to them automatically.  This can be used to register other classes
and functions.

These classes are extracted by static analysis, so the decorator is a noop,
returning the class unchanged.