stbt.Keyboard: Automatically refresh page if it's out of date
This bypasses a common error in user code when we forget to update
`page` between calls to `enter_text` and `navigate_to`. For example:

    def enter_text(self, text):
        kb.enter_text("peppa pig", page=self)
        kb.navigate_to("SEARCH", page=self)  # self.selection is outdated!
        stbt.press_and_wait("OK")

We consider the Page Object instance to be out of date if its frame is
older than the last keypress. This won't reflect changes that aren't
triggered by the remote control (particularly changes triggered by the
passage of time, such as the screensaver activating) but in practice
it'll solve a common bug.

Of course it doesn't fix any other (non-Keyboard) bug caused by
out-of-date Page Object instances. Maybe this will only add to the
confusion ("If this code pattern causes a bug, how come the same pattern
does work when it's using stbt.Keyboard?"). Or maybe this is a good
pattern that we can start using elsewhere. Thoughts please.