stbt.Region: Add `center` property
This is useful when comparing if some region is above another region,
without having to worry about getting the edges of each region exactly
right. e.g. we might have region1 detected at run-time with `stbt.match`
and region2 hard-coded in the test scripts:

    is_above = region1.center.y < region2.y

If we used `region1.y < region2.y` we would get false positives if
region1 *is* at region2 (the selection is on the desired target) but the
way we calculated region1 is off by 1 pixel compared to region2.

If we used `region1.bottom < region2.y` we would get false negatives if
region1 *is* in the position immediately above region2 but, again, we
got the region off by a few pixels (or the seleted item is bigger and
overlaps surrounding items slightly).

Using the center of region1 solves both of these problems.