stbt.load_image: Fix UnicodeDecodeError when filename is utf-8 bytes
`future.utils.text_to_native_str` only works if the input is unicode. It
does this:

    return unicode(t).encode(encoding)

And the `unicode(t)` is what fails.

It seems that we had tested bytes & unicode in
`test_load_image_with_unicode_filename` at some point in the past, but
not since we added `from __future__ import unicode_literals`.

Note that b"Röthlisberger" is a SyntaxError on Python 3: bytes can only
contain ASCII literal characters.

I had to do some contortions with the IOError exceptions we raise. If
you give it a unicode message on Python 2, it displays like this:
"IOError: <unprintable IOError object>".