ImageLogger: Add support for Jupyter Notebooks
If running inside a Jupyter Notebook, display each debug html in an
iframe.

Since we're displaying each one in its own iframe, don't show the "prev"
& "next" links.

Note: Each notebook is run with $PWD set to the directory containing the
notebook (the `.ipynb` file) regardless of where you started the jupyter
server from; so that's where the `stbt-debug` directory is created, and
the relative link in the iframe's `src` attribute works correctly.