stbt.load_image: Return stbt.Image (numpy subclass)
...with additional `filename`, `absolute_filename` and
`relative_filename` attributes.

`relative_filename` is calculated relative to `stbt_core.TEST_PACK_ROOT`
if it exists. You can set it from any code that will run early enough,
for example in `tests/__init__.py`:

    import os
    import stbt_core
    stbt_core.TEST_PACK_ROOT = os.path.abspath(
        os.path.join(os.path.dirname(__file__), ".."))

On the Stb-tester commercial platform `TEST_PACK_ROOT` is automatically
set to the root of the test-pack git repository. In future we may want
`stbt run` to set this, but we'd have to define how to determine the
root -- not all open-source users will be using git repos.