Remove internal _ImageFromUser & _load_image
In favour of the new `stbt.Image` which tracks the image's filename in a
better & more consistent way.

The precise string in some of the debug, exceptions, video annotations,
and MatchResult/MotionResult `image` or `mask` attributes, may have
changed (e.g. "<Image>" instead of "<Custom Image>", or "None" in some
cases) but otherwise this is an internal refactoring.