stbt run: Don't expose `press` etc directly as globals
This code was added for backwards compatibility with a very early `stbt
run` behaviour. It's long overdue for removal. This doesn't affect any
of our customers as they all use the invocation of the test-runner that
runs a specific function in the file, instead of running the whole file
as a script.