HdmiCecControl: Re-scan CEC bus if transmit fails
The implementation is heavily constrained by the fact that we see
frequent `Transmit` errors on the Amazon Fire stick unless you set
stbt's default `interpress_delay_secs` to 5 seconds. Note that
`press_and_hold` requires us to send the keydown command every 300ms.

The semantics are: The code needs to ensure that `self.destination` is
not None before calling `keydown_command()` or `keyup_command()`.
If you call `rescan()` it'll set `self.destination`. If a keypress fails
we set `self.destination = None` and then the next time you try a
keypress we'll rescan the CEC bus.

So in `press` we set `self.destination = None` if the keydown failed,
but we ignore keyup failures (for the purposes of re-scanning; we still
report them by raising an exception) because... I don't know, it seems
unlikely that the device will have disappeared between the keydown and
the keyup? So the keyup failure is more likely to be some other problem?
If it really has gone away then the next `keydown` will fail and we'll
re-scan then.

Similarly, in `keydown` we set `self.destination = None` if the keydown
failed, and in `keyup` and `send_keydowns` we ignore it. Note that in
`send_keydowns` we (already) don't raise exceptions for failures,
because of those intermittent keydown failures we see on the Amazon
Fire.

If you've specified a destination explicitly, it'll still get replaced
by `rescan()` if any transmission fails. I don't know if anyone is
actually using HdmiCecControl with an explicit destination number, or
if this would affect them.

All this is unlikely to affect test scripts, because we still raise an
exception so the test will stop anyway. It'll affect longer-running
processes such as stbt-control-relay.