stbt.Keyboard: Handle transitions on KEY_OK (for shift -> unshifted transitions)
(So far) we don't have any special support for defining "shift" mode
--you just do it with `add_transition`-- but we do need to exclude these
edges when navigating to the next letter; otherwise it'll try to take a
shortcut. For example pressing OK navigates from "A" to "a", but it also
types in a letter.

We also need to refresh the state of the keyboard after pressing OK, to
notice that we're no longer in shifted mode.