stbt.Keyboard: add_grid returns a new Grid of Key objects
This makes it easier to define transitions from the edges of the grid on
to other keys. If you use the input grid, i.e. just the key
names (strings), then it might be ambiguous -- for example the Disney+
search keyboard on Roku has several keys with the same name (in fact the
name is "" because they're blank, but still navigatable). For
add_transition we need to identify the key unambiguously.