Keyboard: Allow using something other than `press_and_wait`
This can be useful for transparent backgrounds, or to implement other means
of interaction.  I've not made it public for now, so we have more freedom
to play, but we can make it more public later if it turns out to be useful,
even without a new stb-tester release.

I've used a base class here rather than duck typing to make the behaviour
more explicitly opt-in.