press_and_wait: "has animation ended?" algorithm detects slow motion
This is the same change we made to `wait_for_motion` in commit
4be487d ("wait_for_motion: Do better at detecting slow motion"):

* While there's no motion, it keeps checking against the original frame.

* If there's motion, it updates "original frame" to be the latest frame,
  so next time it'll be checking against the immediately previous frame.
  This allows it to reliable detect the transition to "no motion".

Note that the "wait for animation to start" part of press_and_wait
already behaved like this.

The real reason I'm doing this is for consistency with
`wait_for_motion`, so that I'll be able to refactor the `strict_diff`
algorithm out into a `FrameDiffer` class.