press_and_wait unit tests: Test with MotionDiff too
`diff_algorithm` is a "parametrized fixture". Any test that depends on
it will be called multiple times. The test doesn't need to be aware of
this. See https://docs.pytest.org/en/stable/fixture.html#fixture-parametrize

In `FakeDeviceUnderTest` I had to make the circle filled instead of
hollow, because the thin border was entirely removed MotionDiff's erode
step.