Define MotionResult.__nonzero__ for Python 2
When I move MotionResult from motion.py to the new file diff.py, I
forgot to add the litany of Python3 compatibility at the top. Normally
we say `from builtins import *` (provided by python-future) which
changes the `object` base class so that it automatically provides a
`__nonzero__` if you provide a `__bool__`. But let's just define it
ourselves and see how we get on -- it's a small first step to stop
depending on python-future which has proved problematic.