press_and_wait: Allow specifying both mask and region
`wait_for_motion`, `detect_motion` and `is_screen_black` all allow
specifying region *and* mask since we added the region parameter in v28.
If you specify both `region` and `mask`, the mask must be the same size
as the region. In other words, we apply `region` first and then `mask`.