stbt.ocr: Enable caching by default
This will free us to use OCR more, well, freely, in menu navigation,
`wait_until` loops, etc.

The cache uses little space and hashing the frame is very fast (thanks
to libxxhash) so there's little reason not to enable it, and we don't
expect a need to disable it. In extreme need you can set
`_stbt.imgproc_cache._cache = None`.