stbt.Keyboard: Add press_and_wait member
So that callers can override it if they need to provide a custom
implementation (for example to handle live video behind a translucent
background). I won't document this for now, but it gives us some
flexibility without having to make a new Stb-tester release.

This is an interim measure until we iron out the "Movable" API in
<https://github.com/stb-tester/stb-tester/pull/695>.