stbt.press_and_wait: Add min_size parameter
When testing an Xfinity set-top-box, I think it was, pressing OK on a
button caused that button to "blip" briefly for a single frame or 2,
then followed by no motion, before actually starting the transition to
the new page. Sometimes that period of no motion was longer than 1
second, so it would fool `press_and_wait` into thinking the transition
had completed before it had even started.

Setting `stable_secs` to a very large number isn't ideal, as it would
slow down all test-runs for the sake of handling this intermittent
behaviour.

Instead, we can use this new `min_size` parameter so that it ignores
the blip on the button, and looks for changes to the whole screen.

Other use-cases:

- Ignore the scrolling text below the selected tile in the Apple TV's
  app store.
- Ignore the blinking cursor in a search box (though this one is
  probably already ignored by MotionDiff's erode step).

At first I wasn't sure if `min_size` should affect the decision whether
to update `previous_frame`. I decided that it *should* affect the
decision -- that is, if we found motion but it's still smaller than
min_size, we treat it exactly the same as "no motion" and we don't
update `previous_frame`. That way, if something is moving along slowly,
the differences will eventually exceed `min_size` and count as "motion".
This way the algorithm will eventually detect motion regardless of the
frame-rate it's running at.

In this way `min_size` can mask out a static area of a particular size,
but not a *moving* area.