match/ocr/etc: Raise exception if region is None or outside of frame
For `ocr` and `match_text` this is a change in behaviour: Previously if
the region was entirely outside of the frame, they would log a warning
and return "" or a Falsey TextMatchResult, respectively (but for
`region=None` they would already raise a TypeError). I don't think the
old behaviour is anything you'd ever want to do intentionally, it's far
more likely to be a bug in your test. And the new behaviour is
consistent with all the other stbt APIs.

For all the other functions this is a nicer error message.
`press_and_wait` would raise "UnboundLocalError: local variable 'f1'
referenced before assignment". The others would raise either
"ValueError: frame with dimensions (720, 1280, 3) doesn't contain
None" (from stbt.crop) or "AttributeError: 'NoneType' object has no
attribute 'x/y/right/bottom'".