imgproc_cache: Add use_cache parameter to memoized functions
This allows enabling the cache for individual functions in a thread-safe
way. The previous way I was enabling caching for `_tesseract_subprocess`
was not thread safe: Another thread doing a `stbt.match` might suddenly
get cached too.

We still need the `enable_caching` contextmanager to enable caching
globally, for example for stbt_auto_selftest.py and the Object
Repository.

Note: I have to do `kwargs.pop()` because the following is a SyntaxError
on Python 2.7:

    def inner(*args, use_cache=None, **kwargs):
        ...

Unfortunately this confuses pylint. If we ever drop support for Python
2.7 we can remove the `# pylint:disable=unexpected-keyword-arg`.