CI: python-package job: Upgrade pytest to 4.6.9
This is the latest pytest version that still supports Python 2. Though
here we're testing with Python 3. Hmmm it might be worth upgrading
higher still but at least this fixes my immediate problem:

I added the latest version of `attrs` as a dependency to setup.py (in
the previous commit) but this new version of `attrs` doesn't work with
our old version of pytest (our old pytest was passing a parameter to
`attr.ib` that is no longer supported).