stbt.Keyboard: Wait longer for selection to reach the target
I was testing the Search keyboard on a Comcast Xfinity box. This has the
keys in a single line A - Z so there can be a lot of keypresses to get
to the next letter. When `verify_every_keypress=False`, the selection
can settle on one of the intermediate letters for long enough to trick
our `stable_secs=0.5`, causing the test to fail. After the test failed,
the selection carried on moving and did reach the expected target.

I tried increasing stable_secs to 1. This made it fail less often, but
it still happens sometimes. Setting `verify_every_keypress=True` fixes
this problem but makes it intolerably slow. I don't want to increase
stable_secs for every letter because this would slow it down
unnecessarily in the usual, successful case.