Fix importing stbt_core when libstbt.so isn't built/installed
In 0eb1f63 ("stbt.match: Make the sqdiff C optimisation optional") we catch
ImportError when importing _stbt.sqdiff, so that we could publish a
pure-python package to PyPI without having to worry about C compilation or
binary wheels. However it actually raises OSError:

    tests/test_match.py:17: in <module>
        import stbt_core as stbt
    stbt_core/__init__.py:44: in <module>
        from _stbt.match import (
    _stbt/match.py:34: in <module>
        from .sqdiff import sqdiff
    _stbt/sqdiff.py:18: in <module>
        _libstbt = ctypes.CDLL(_find_file("libstbt.so"))
    /usr/lib/python3.6/ctypes/__init__.py:348: in __init__
        self._handle = _dlopen(self._name, mode)
    E   OSError: /home/drothlis/work/stb-tester.com/stb-tester/_stbt/libstbt.so: cannot open shared object file: No such file or directory

Instead here we only load `libstbt.so` on first use.  This means that we
can `import stbt_core` without `libstbt.so`, but we will still get an error
if we actually do need to use it.

I didn't want to automatically fall-back to our numpy implementation as
this may mask real installation issues.