load_image: Replace `flags` parameter with `color_channels`
This is a bit more explicit about exactly what shape the output image will
be.  We now have the simple post-condition:

    out.shape[2] in color_channels

whereas previously if we passed an existing image as `filename` it would
be returned unmodified.  I have hit issues where I've tried to pre-load a
mask, but then it has the wrong number of colour channels for the function
I pass it to.

We also remove (deprecate) the dependency on `cv2` from our interface which
is poorly documented in the opencv documentation.  This means that
customers who use `load_image` no longer need to `import cv2`.