TransitionResult: Add started/complete/stable properties
This allows us to write:

    transition = stbt.press_and_wait(...)
    assert transition.stable

or even this:

    assert stbt.press_and_wait(...).stable

instead of this, which is cumbersome:

    assert transition.status != stbt.TransitionStatus.STABLE_TIMEOUT

See commit bc46fcc399fa ("stbt.Keyboard: Use stable frame for
page.refresh()") for an example where we need to check
`transition.status != stbt.TransitionStatus.STABLE_TIMEOUT`.

Another case is:

    if not transition.started:
        # reached the end of the menu
        ...

...to detect if we've reached the end of a menu because the keypress
had no effect. `if not transition` isn't suitable because it could mean
not started *or* started but not completed.