stbt.Image: Add `width` and `height` properties
Recently I had a reference image that I wanted to anchor to the left
edge of the frame like this:

    image = stbt.load_image(...)
    stbt.match(image, region=stbt.Region(x=0, width=image.shape[1], ...))

...so that if I had to update the reference image in future I wouldn't
have to update the Python code to match the new width.

By providing `width` and `height` properties, we don't have to expose
the numpy details of `shape` and its unintuitive coordinate order.