stbt.Frame: Add `width` and `height` properties
Similar to the recent commit 656a59f2d3a4 "stbt.Image: Add `width` and
`height` properties", but for `stbt.Frame`.

Recently I was checking for the width of a region obtained with
`right_of`. But this region has infinite width, so I needed to intersect
it with the size of the frame, but this isn't currently possible without
reaching into `shape` which is numpy-specific and has unintuitive
coordinate order. Now I can do something like
`r.right_of().replace(right=self._frame.right)`.

Note that the current documentation of `right_of` is incorrect, or at
least misleading: "Returns: A new region to the right of the current
region, extending **to the right edge of the frame** (or to the
specified width)" -- emphasis mine.