test_that_video_index_is_written_on_eos: Test by playing an mp4 file
This removes the dependency on webminspector.py which hasn't been ported
to Python3, so the test was being skipped in our CI job.

Also, webminspector.py didn't like the webm file created by `stbt run`
on Ubuntu 20.04 (something changed in GStreamer?) so even if I ran the
test locally with Python 2 installed, the test would fail.

MP4 isn't streamable, so we can test that the EOS was sent & the MP4
file was written out completely, by trying to play the file. I tested
this by commenting out the `self.appsrc.emit("end-of-stream")` line
in `_stbt/core.py`, and the gst-launch in the test fails with:

> ERROR: from element /GstPipeline:pipeline0/GstQTDemux:qtdemux0: This file contains no playable streams.
> Additional debug info:
> qtdemux.c(716): gst_qtdemux_post_no_playable_stream_error (): /GstPipeline:pipeline0/GstQTDemux:qtdemux0:
> no known streams found
> ERROR: pipeline doesn't want to preroll.