stbt.press, stbt.pressing: Convert Enum to string
`str(key)` doesn't work because it returns `"Keys.KEY_MENU"` but we need
`"KEY_MENU"`. So we have to check that it's an Enum explicitly.

I have added a test too (which is how I noticed this problem).

It's nice that we don't have to worry about Python 2.7 (which doesn't
have `enum` in its standard library) any more.