stbt.ocr corrections: Match non-word characters at string boundaries
When you specify a plain string as an OCR correction, it's supposed to
match at a word boundary (if you want anything more complex, you can
specify a compiled regular expression). `re` defines word boundaries as
going from a word character to a non-word character (the start or end of
the string count as non-word characters). But if your correction string
starts or ends with a non-word character, we don't consider the
transition from that (non-word) character, to another non-word
character, as a match.

The goal of this logic is to replace complete words or phrases without
the surprising behaviour of replacing sub-strings inside another longer
word. So I have re-defined "word boundary" to mean whitespace or the end
of the string. I think that achieves our goal, while also fixing this
bug.

This whole feature was severely under-tested. I have added some other
test-cases, too, that caught bugs in my initial attempts at this fix.