stbt.get_frame: Discard frames older than the last keypress
Typically you want to see the effect of the keypress, something like
this:

    stbt.press("KEY_RIGHT")
    stbt.wait_until(lambda: stbt.match(...))

You don't want to waste time performing expensive operations on a frame
that's in the video-capture pipeline but pre-dates the keypress.

This change affects calls to `get_frame` or `frames` immediately after a
keypress. It doesn't affect the next frames from a `frames` generator
started before the call to `press`. So in the following loop, you'll
potentially get frames from before the press:

    frames = stbt.frames()
    f = next(frames)
    stbt.press("KEY_RIGHT")
    for f in frames:
        ...