stbt.Keyboard: Better error message when page.selection is None
If your Page Object is buggy, and it returns `is_visible=True` but
`selection=None` we'd get an inscrutable TypeError from the depths of
our implementation:

>   File "_stbt/keyboard.py", line 707, in _keys_to_press
>     for s, t in zip(path[:-1], path[1:]):
> TypeError: unhashable type: 'slice'

This is because we call `networkx.shortest_path(G, source, target)`.
Normally it returns a list, but when `source` is None it returns a
dict (with the shortest path for each node in the graph to `target`,
keyed by the starting node).