[WIP] Fix stbt-debug rendering in Jupyter Notebooks
With Jupyter Notebook 6.4.6 running on Ubuntu 20.04, the IFrame with the
stbt-debug index.html doesn't show any of the images. Looking in the
network tab of my browser's developer tools, I can see that the requests
for the images are being redirected to
"login?next=<path to the actual image.png>".

If I right-click the broken image and open it in a new tab, it works;
but it still doesn't display the image in the HTML.

This commit renders correctly in Jupyter but obviously it's a hack
because it breaks the normal index.html when viewed directly.

TODO: Fix it properly.