Remove useless pylint annotations
I found these by running pylint with `--enable=useless-suppression`.

I learned of this in the pylint documentation:
https://pylint.pycqa.org/en/latest/user_guide/message-control.html#detecting-useless-disables

> As pylint gets better and false positives are removed, disables that
> became useless can accumulate and clutter the code. In order to clean
> them you can enable the useless-suppression warning.

I haven't enabled `useless-suppression` permanently in pylint.conf
because it says that the `line-too-long` annotations in
`tests/test_keyboard.py` are useless, but if I remove the annotations
then pylint complains that the lines are too long.

Tested with Ubuntu 20.04's pylint which is what we use in CI:

    ❯ pylint --version
    pylint 2.4.4
    astroid 2.3.3