Add test_that_partly_opaque_alpha_is_treated_as_fully_transparent
To check that this code in `_find_matches` is necessary:

    if template.shape[2] == 4:
        # Normalise transparency channel to either 0 or 255
        mask = template[:, :, 3]
        mask[mask < 255] = 0

Note that OpenCV does its own thresholding if the alpha channel is U8
(which it always is, in our code) but OpenCV treats a partly-transparent
pixel as fully opaque, whereas we set it to fully transparent.

`button-partly-transparent.png` is the same as `button-transparent.png`
but with transparent pixels in the alpha channel set to 128 instead of
0, with this code:

    f = stbt.load_image("tests/button-transparent.png")
    mask = f[:,:,3]
    mask[mask == 0] = 128
    cv2.imwrite("tests/button-partly-transparent.png", f)

I checked that the new test fails if I comment out the above code in
`_find_matches`.