load_image: Cache reading PNG from disk
If you're calling `match("reference.png")` in a loop with `wait_until`
we don't need to decode the PNG afresh each time.

The cache size is small because each image can take several megabytes,
but larger than 1 because sometimes we do `match(a) or match(b)` and
also to support operations that load a reference image and a mask.