Mask: Allow adding/subtracting empty regions
`None` means an empty region, for example the result of
`Region.intersect` if the input regions don't overlap.

We can handle `None` on either side of the operator, by implementing
the reflected operators (`__radd__` and `__rsub__`). See
https://docs.python.org/3/reference/datamodel.html#object.__radd__

We still can't support `None + None` but that's a limitation I can live
with for now. Fixing this properly will require adding a new singleton
to represent empty regions, changing `Region.intersect` to return that
instead of `None`, and updating the implementation of all our other
region-handling code. There's a (out of date) pull request for that:
https://github.com/stb-tester/stb-tester/pull/624