integration tests: timeout with SIGINT
Hopefully this will give us a python stack trace when the stbt-run
command times out. This is to debug test failures on CircleCI.
`test_that_invalid_control_doesnt_hang` failed, and it seems ludicrous
that this would take longer than 20 seconds, even on CircleCI. Maybe
there is a real bug?