stbt-run: Test formatting of IOError exceptions
This is a regression test for incorrect behaviour in our (now removed)
Python 2 implementation. We don't need any code changes now that we're
not using Python 2 / python-kitchen, but the tests are worth keeping.

Previously code like:

    open('non-existent')

would print the error:

    IOError: '2'

because 2 is `errno.ENOENT`.  Now it prints:

    IOError: [Errno 2] No such file or directory: 'non-existent'

which is the same that Python would print.