CI: Rename job to "make_check-pythonpackage"
So that it doesn't have the same name as the job in the "ubuntu2204"
workflow.

I'm trying to configure the "require status checks to pass before
merging" but the search-box for "status checks in the last week for this
repository" is only returning matches for "make_check" (the name of the
jobs), not for "python-package" or "ubuntu2204" (the name of the
workflows).