MotionDiff: Rename noise_threshold -> threshold
This is just the MotionDiff parameter; the detect_motion/wait_for_motion
parameter is still called "noise_threshold".

It was always a bad name but I don't want to change it for
detect_motion/wait_for_motion, for backward compatibility.
MotionDiff is technically a public API but it isn't documented and
no-one uses it directly, as far as I know.