match, ocr: imgproc_cache cache-buster for v33
In case there are changes in behaviour due to new Tesseract / OpenCV
versions.

Regarding stbt.match: In spite of the new `matchTemplate` implementation
in OpenCV 4.4, we have extensive tests that check the match position,
and none of those tests have required changes. Maybe the
`first_pass_result` could vary very slightly, resulting in a change in
behaviour when it was already very close to the threshold?