pylint_plugin: Increase astroid's inference limit
Pylint has `--limit-inference-results` (defaults to 100) but it only
affects the limit per infer() call. This `InferenceContext.max_inferred`
is a limit to the whole InferenceContext (whatever a context is -- maybe
per file?) and it's hard-coded to 100.

This would have fixed some of the no-member false positives I was
seeing, but the newer version of pylint (see previous commit) also fixes
them. Still, having better inference results is a good thing (once you
hit this limit, infer() starts returning `Uninferable`).