debug: Use python's logging framework
Note: This leaves the root logger unconfigured, so that we don't change
the behaviour of users' existing calls to `logging.basicConfig` (because
`basicConfig` is ignored if you call it more than once, or if you've
already added handlers to the root logger some other way).

We set `propagate = False` so that our logger's messages don't get
printed again by the root logger.

Eventually we'll make all of stb-tester's code call the logging
framework directly —e.g. `logging.getLogger("stbt.match").debug(...)`—
but for now this is an easy step forward.