logging: Use Python's default logging format
For example DEBUG:stbt.Keyboard:This is the message...

This is the default format if you call `logging.basicConfig()` without
specifying a format (though it's not the default format when you create
a Formatter explicitly: see <https://stackoverflow.com/a/72801269>).

This is the (undocumented) `logging.BASIC_FORMAT`.

I'm changing it so that it's more likely to match the user's own
logging. If you call `logging.info` (or debug etc) it'll automatically
call `logging.basicConfig()`.