Mask.to_array: Take region & color_channels instead of shape
For now, at least, this region represents the size of the full frame —
we're not cropping inside the mask, so we expect the region to be
anchored at 0, 0. You can't specify Region.ALL either; we need to know
the size of the frame.

The current documentation (and behaviour) of detect_motion,
is_screen_black, and press_and_wait says:

> If you specify ``region``, the mask must be the same size as the
> region. Otherwise the mask must be the same size as the frame.

This commit preserves that behaviour. It might make sense to change
that, or to disallow specifying `region` and `mask` simultaneously.

Notes:

- If you ask for color_channels=3 we're also memoizing color_channels=1
  because it'll be useful for a future implementation of
  `Mask.bounding_box`.

- On my laptop `cv2.cvtColor(array, cv2.COLOR_GRAY2BGR)` was 2-3x faster
  than `numpy.concatenate([array, array, array], axis=2)` (400-700µs vs.
  1.3ms) or 9x faster when run in a loop with ipython's `%timeit`
  (120µ vs. 1.1ms).